Refactor GUI: extract widgets and mixins
