+++ codes/mainwindow.py
@@ -45,6 +45,10 @@ from workers.DEWorker import DEWorker
 from workers.SAWorker import SAWorker
 from workers.CMAESWorker import CMAESWorker
 # RL module import removed
+from gui.widgets import ModernQTabWidget, SidebarButton
+from gui.menu_mixin import MenuMixin
+from gui.beam_mixin import ContinuousBeamMixin
+from gui.microchip_mixin import MicrochipPageMixin
 
 
 # Additional libraries used
@@ -73,46 +77,9 @@ except ImportError:
     print("Warning: Continuous Beam module imports failed. This functionality will be disabled.")
 
 
-class ModernQTabWidget(QTabWidget):
-    """Custom TabWidget with modern styling"""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setDocumentMode(True)
-        self.setTabPosition(QTabWidget.North)
-        self.setMovable(True)
-
-
-class SidebarButton(QWidget):
-    """Custom styled sidebar button"""
-    def __init__(self, icon_path, text, parent=None):
-        super().__init__(parent)
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(10, 5, 10, 5)
-        
-        if icon_path:
-            icon = QLabel()
-            pixmap = QPixmap(icon_path)
-            icon.setPixmap(pixmap.scaled(24, 24, Qt.KeepAspectRatio, Qt.SmoothTransformation))
-            layout.addWidget(icon)
-        
-        label = QLabel(text)
-        label.setFont(QFont("Segoe UI", 10))
-        layout.addWidget(label)
-        layout.addStretch()
-        
-        self.setMinimumHeight(50)
-        self.setCursor(QCursor(Qt.PointingHandCursor))
-    
-    def enterEvent(self, event):
-        self.setStyleSheet("background-color: rgba(255, 255, 255, 0.2); border-radius: 5px;")
-        super().enterEvent(event)
-    
-    def leaveEvent(self, event):
-        self.setStyleSheet("")
-        super().leaveEvent(event)
         
 
-class MainWindow(QMainWindow):
+class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin):
     def __init__(self):
         super().__init__()
         self.setWindowTitle("DeVana")
@@ -7397,318 +7364,6 @@ class MainWindow(QMainWindow):
         self.status_bar.showMessage("Reset to default values")
         # Reset logic for all parameters would be implemented here
 
-    def create_menubar(self):
-        """Create the application menubar with modern styling"""
-        menubar = self.menuBar()
-        
-        # File menu
-        file_menu = menubar.addMenu("&File")
-        
-        # New Project
-        new_action = QAction("&New Project", self)
-        new_action.setShortcut("Ctrl+N")
-        new_action.triggered.connect(lambda: self.status_bar.showMessage("New Project - Feature coming soon"))
-        file_menu.addAction(new_action)
-        
-        # Open Project
-        open_action = QAction("&Open Project", self)
-        open_action.setShortcut("Ctrl+O")
-        open_action.triggered.connect(lambda: self.status_bar.showMessage("Open Project - Feature coming soon"))
-        file_menu.addAction(open_action)
-        
-        file_menu.addSeparator()
-        
-        # Save Project
-        save_action = QAction("&Save Project", self)
-        save_action.setShortcut("Ctrl+S")
-        save_action.triggered.connect(lambda: self.status_bar.showMessage("Save Project - Feature coming soon"))
-        file_menu.addAction(save_action)
-        
-        # Save Project As
-        save_as_action = QAction("Save Project &As", self)
-        save_as_action.setShortcut("Ctrl+Shift+S")
-        save_as_action.triggered.connect(lambda: self.status_bar.showMessage("Save Project As - Feature coming soon"))
-        file_menu.addAction(save_as_action)
-        
-        file_menu.addSeparator()
-        
-        # Import
-        import_action = QAction("&Import Parameters", self)
-        import_action.triggered.connect(self.import_parameters)
-        file_menu.addAction(import_action)
-        
-        # Export
-        export_action = QAction("&Export Parameters", self)
-        export_action.triggered.connect(self.export_parameters)
-        file_menu.addAction(export_action)
-        
-        file_menu.addSeparator()
-        
-        # Exit
-        exit_action = QAction("E&xit", self)
-        exit_action.setShortcut("Alt+F4")
-        exit_action.triggered.connect(self.close)
-        file_menu.addAction(exit_action)
-        
-        # Edit menu
-        edit_menu = menubar.addMenu("&Edit")
-        
-        # Default values
-        default_action = QAction("Reset to &Default Values", self)
-        default_action.triggered.connect(self.set_default_values)
-        edit_menu.addAction(default_action)
-        
-        # View menu
-        view_menu = menubar.addMenu("&View")
-        
-        # Toggle Sidebar
-        sidebar_action = QAction("Toggle &Sidebar", self)
-        sidebar_action.setShortcut("Ctrl+B")
-        sidebar_action.triggered.connect(lambda: self.status_bar.showMessage("Toggle Sidebar - Feature coming soon"))
-        view_menu.addAction(sidebar_action)
-        
-        view_menu.addSeparator()
-        
-        # Theme submenu
-        theme_menu = view_menu.addMenu("&Theme")
-        
-        # Theme action group to make selections exclusive
-        theme_group = QActionGroup(self)
-        
-        # Dark theme action
-        dark_action = QAction("&Dark Theme", self)
-        dark_action.setCheckable(True)
-        if self.current_theme == 'Dark':
-            dark_action.setChecked(True)
-        dark_action.triggered.connect(lambda: self.switch_theme('Dark'))
-        theme_group.addAction(dark_action)
-        theme_menu.addAction(dark_action)
-        
-        # Light theme action
-        light_action = QAction("&Light Theme", self)
-        light_action.setCheckable(True)
-        if self.current_theme == 'Light':
-            light_action.setChecked(True)
-        light_action.triggered.connect(lambda: self.switch_theme('Light'))
-        theme_group.addAction(light_action)
-        theme_menu.addAction(light_action)
-        
-        # Tools menu
-        tools_menu = menubar.addMenu("&Tools")
-        
-        # Run FRF
-        run_frf_action = QAction("Run &FRF Analysis", self)
-        run_frf_action.triggered.connect(self.run_frf)
-        tools_menu.addAction(run_frf_action)
-        
-        # Run Sobol
-        run_sobol_action = QAction("Run &Sobol Analysis", self)
-        run_sobol_action.setToolTip("Run Sobol Analysis")
-        run_sobol_action.setVisible(False)  # Hide button
-        tools_menu.addAction(run_sobol_action)
-        
-        tools_menu.addSeparator()
-        
-        # Optimization submenu
-        optimization_menu = tools_menu.addMenu("&Optimization")
-        
-        # GA Optimization
-        ga_action = QAction("&Genetic Algorithm", self)
-        ga_action.triggered.connect(self.run_ga)
-        optimization_menu.addAction(ga_action)
-        
-        # PSO Optimization
-        pso_action = QAction("&Particle Swarm", self)
-        pso_action.triggered.connect(self.run_pso)
-        optimization_menu.addAction(pso_action)
-        
-        # DE Optimization
-        de_action = QAction("&Differential Evolution", self)
-        de_action.triggered.connect(self.run_de)
-        optimization_menu.addAction(de_action)
-        
-        # SA Optimization
-        sa_action = QAction("&Simulated Annealing", self)
-        sa_action.triggered.connect(self.run_sa)
-        optimization_menu.addAction(sa_action)
-        
-        # CMAES Optimization
-        cmaes_action = QAction("&CMA-ES", self)
-        cmaes_action.triggered.connect(self.run_cmaes)
-        optimization_menu.addAction(cmaes_action)
-        
-        # RL Optimization removed
-        
-        # Help menu
-        help_menu = menubar.addMenu("&Help")
-        
-        # Documentation
-        docs_action = QAction("&Documentation", self)
-        docs_action.triggered.connect(lambda: self.status_bar.showMessage("Documentation - Feature coming soon"))
-        help_menu.addAction(docs_action)
-        
-        # About
-        about_action = QAction("&About", self)
-        about_action.triggered.connect(self.show_about)
-        help_menu.addAction(about_action)
-    
-    def show_about(self):
-        """Show about dialog"""
-        QMessageBox.about(self, "About DeVana", 
-                          "DeVana v2.0\n\n"
-                          "A modern application for designing and optimizing vibration systems.\n\n"
-                          "© 2023 DeVana Team\n"
-                          "All rights reserved.")
-        
-    def create_toolbar(self):
-        """Create the application toolbar with modern styling"""
-        toolbar = QToolBar("Main Toolbar")
-        toolbar.setMovable(False)
-        toolbar.setIconSize(QSize(24, 24))
-        self.addToolBar(toolbar)
-        
-        # Add spacer at the beginning
-        spacer = QWidget()
-        spacer.setFixedWidth(10)
-        toolbar.addWidget(spacer)
-        
-        # New Project button
-        new_button = QPushButton("New Project")
-        new_button.setObjectName("toolbar-button")
-        new_button.setToolTip("Create a new project")
-        new_button.clicked.connect(lambda: self.status_bar.showMessage("New Project - Feature coming soon"))
-        toolbar.addWidget(new_button)
-        
-        # Open Project button
-        open_button = QPushButton("Open Project")
-        open_button.setObjectName("toolbar-button")
-        open_button.setToolTip("Open an existing project")
-        open_button.clicked.connect(lambda: self.status_bar.showMessage("Open Project - Feature coming soon"))
-        toolbar.addWidget(open_button)
-        
-        # Save Project button
-        save_button = QPushButton("Save Project")
-        save_button.setObjectName("toolbar-button")
-        save_button.setToolTip("Save the current project")
-        save_button.clicked.connect(lambda: self.status_bar.showMessage("Save Project - Feature coming soon"))
-        toolbar.addWidget(save_button)
-        
-        # Add separator
-        toolbar.addSeparator()
-        
-        # Run FRF button
-        run_frf_button = QPushButton("Run FRF")
-        run_frf_button.setObjectName("primary-button")
-        run_frf_button.setToolTip("Run FRF Analysis")
-        run_frf_button.clicked.connect(self.run_frf)
-        run_frf_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_frf_button)
-        
-        # Run Sobol button
-        run_sobol_button = QPushButton("Run Sobol")
-        run_sobol_button.setObjectName("primary-button")
-        run_sobol_button.setToolTip("Run Sobol Analysis")
-        run_sobol_button.clicked.connect(self._run_sobol_implementation)
-        run_sobol_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_sobol_button)
-        
-        # Run PSO button
-        run_pso_button = QPushButton("Run PSO")
-        run_pso_button.setObjectName("primary-button")
-        run_pso_button.setToolTip("Run Particle Swarm Optimization")
-        run_pso_button.clicked.connect(self.run_pso)
-        run_pso_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_pso_button)
-        
-        # Run DE button
-        run_de_button = QPushButton("Run DE")
-        run_de_button.setObjectName("primary-button")
-        run_de_button.setToolTip("Run Differential Evolution")
-        run_de_button.clicked.connect(self.run_de)
-        run_de_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_de_button)
-        
-        # Run SA button
-        run_sa_button = QPushButton("Run SA")
-        run_sa_button.setObjectName("primary-button")
-        run_sa_button.setToolTip("Run Simulated Annealing")
-        run_sa_button.clicked.connect(self.run_sa)
-        run_sa_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_sa_button)
-        
-        # Run CMA-ES button
-        run_cmaes_button = QPushButton("Run CMA-ES")
-        run_cmaes_button.setObjectName("primary-button")
-        run_cmaes_button.setToolTip("Run CMA-ES Optimization")
-        run_cmaes_button.clicked.connect(self.run_cmaes)
-        run_cmaes_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_cmaes_button)
-        
-        # Add separator
-        toolbar.addSeparator()
-        
-        # Theme toggle
-        theme_button = QPushButton("Toggle Theme")
-        theme_button.setObjectName("toolbar-button")
-        theme_button.setToolTip(f"Switch to {'Light' if self.current_theme == 'Dark' else 'Dark'} Theme")
-        theme_button.clicked.connect(self.toggle_theme)
-        toolbar.addWidget(theme_button)
-        
-        # Add expanding spacer
-        spacer = QWidget()
-        spacer.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Preferred)
-        toolbar.addWidget(spacer)
-        
-    def switch_theme(self, theme):
-        """Switch the application theme"""
-        self.current_theme = theme
-        if theme == 'Dark':
-            self.apply_dark_theme()
-        else:
-            self.apply_light_theme()
-        
-        # Update theme toggle button tooltip
-        for action in self.findChildren(QAction):
-            if action.text() == "Toggle &Theme":
-                action.setToolTip(f"Switch to {'Light' if theme == 'Dark' else 'Dark'} Theme")
-
-
-    def create_microchip_controller_page(self):
-        """Create the empty microchip controller page for future implementation"""
-        microchip_page = QWidget()
-        layout = QVBoxLayout(microchip_page)
-        
-        # Centered content
-        center_widget = QWidget()
-        center_layout = QVBoxLayout(center_widget)
-        center_layout.setAlignment(Qt.AlignCenter)
-        
-        # Icon or image placeholder
-        placeholder = QLabel()
-        placeholder.setPixmap(QPixmap("placeholder_image.png" if os.path.exists("placeholder_image.png") else ""))
-        placeholder.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(placeholder)
-        
-        # Title
-        title = QLabel("Microchip Controller")
-        title.setFont(QFont("Segoe UI", 24, QFont.Bold))
-        title.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(title)
-        
-        # Description
-        description = QLabel("This feature will provide interfaces for microchip-based vibration control systems.")
-        description.setFont(QFont("Segoe UI", 12))
-        description.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(description)
-        
-        # Coming soon label
-        coming_soon = QLabel("Coming Soon!")
-        coming_soon.setFont(QFont("Segoe UI", 14, QFont.Bold))
-        coming_soon.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(coming_soon)
-        
-        layout.addWidget(center_widget)
-        self.content_stack.addWidget(microchip_page)
 
     def run_ga(self):
         """Run genetic algorithm optimization"""
@@ -10338,1211 +9993,4 @@ class MainWindow(QMainWindow):
         
         # Add scroll area to the tab's layout
         layout.addWidget(scroll_area)
-
-    def create_continuous_beam_page(self):
-        """Create the continuous beam analysis page"""
-        if not BEAM_IMPORTS_SUCCESSFUL:
-            # Create placeholder page if imports failed
-            beam_page = QWidget()
-            layout = QVBoxLayout(beam_page)
-            
-            # Centered content
-            center_widget = QWidget()
-            center_layout = QVBoxLayout(center_widget)
-            center_layout.setAlignment(Qt.AlignCenter)
-            
-            # Error message
-            error_label = QLabel("Continuous Beam Module Not Available")
-            error_label.setFont(QFont("Segoe UI", 24, QFont.Bold))
-            error_label.setAlignment(Qt.AlignCenter)
-            center_layout.addWidget(error_label)
-            
-            description = QLabel("Please make sure the 'Continues beam' module is correctly installed.")
-            description.setFont(QFont("Segoe UI", 12))
-            description.setAlignment(Qt.AlignCenter)
-            center_layout.addWidget(description)
-            
-            layout.addWidget(center_widget)
-            self.content_stack.addWidget(beam_page)
-            return
-            
-        # Create tab container
-        beam_page = QWidget()
-        page_layout = QVBoxLayout(beam_page)
-        page_layout.setContentsMargins(10, 10, 10, 10)
-        page_layout.setSpacing(10)
-        
-        # Header - more compact
-        header = QWidget()
-        header_layout = QHBoxLayout(header)
-        header_layout.setContentsMargins(5, 5, 5, 5)
-        
-        title_container = QVBoxLayout()
-        title = QLabel("Continuous Beam Analysis")
-        title.setFont(QFont("Segoe UI", 16, QFont.Bold))
-        title_container.addWidget(title)
-        
-        description = QLabel("Analyze and optimize vibration in continuous beams")
-        description.setFont(QFont("Segoe UI", 10))
-        title_container.addWidget(description)
-        
-        header_layout.addLayout(title_container)
-        header_layout.addStretch()
-        
-        # Add header to page layout with minimal height
-        header.setMaximumHeight(70)
-        page_layout.addWidget(header)
-        
-        # Initialize the beam specific properties
-        self.layers = []
-        self.force_regions_manager = ForceRegionManager()
-        
-        # Add a default layer
-        default_layer = {
-            'height': 0.05,
-            'E': 210e9,  # Steel
-            'rho': 7800
-        }
-        self.layers.append(default_layer)
-        
-        # Pre-initialize canvases to avoid NoneType errors
-        self.beam_anim = None
-        self.beam_canvas = None
-        self.node_anim = None
-        self.node_canvas = None
-        
-        # Create tabs
-        self.beam_tabs = ModernQTabWidget()
-        
-        # Add tabs
-        self.init_beam_parameters_tab()
-        self.init_layers_tab()
-        self.init_loads_tab()
-        self.init_beam_results_tab()
-        
-        # Add tabs to the page
-        page_layout.addWidget(self.beam_tabs)
-        
-        # Create button area
-        button_container = QWidget()
-        button_layout = QHBoxLayout(button_container)
-        button_layout.setContentsMargins(0, 10, 0, 0)
-        
-        # Add run button
-        self.run_beam_button = QPushButton("▶ Run Simulation")
-        self.run_beam_button.clicked.connect(self.run_beam_simulation)
-        self.run_beam_button.setMinimumHeight(40)
-        self.run_beam_button.setStyleSheet("""
-            QPushButton {
-                background-color: #4CAF50;
-                color: white;
-                font-weight: bold;
-                border-radius: 4px;
-                padding: 8px 16px;
-            }
-            QPushButton:hover {
-                background-color: #45a049;
-            }
-            QPushButton:pressed {
-                background-color: #388E3C;
-            }
-        """)
-        
-        # Create spacer to push buttons to the right
-        button_layout.addStretch()
-        button_layout.addWidget(self.run_beam_button)
-        
-        # Add button container to main layout
-        page_layout.addWidget(button_container)
-        
-        # Add the page to the stack
-        self.content_stack.addWidget(beam_page)
-        
-    def init_beam_parameters_tab(self):
-        """Initialize the beam parameters tab"""
-        beam_tab = QWidget()
-        beam_layout = QVBoxLayout(beam_tab)
-        beam_layout.setContentsMargins(10, 10, 10, 10)
-        beam_layout.setSpacing(10)
-        
-        # Material properties group
-        material_group = QGroupBox("Material Properties")
-        material_layout = QFormLayout(material_group)
-        material_layout.setFieldGrowthPolicy(QFormLayout.AllNonFixedFieldsGrow)
-        material_layout.setContentsMargins(10, 20, 10, 10)
-        material_layout.setSpacing(10)
-        
-        # Young's modulus
-        self.young_modulus = QDoubleSpinBox()
-        self.young_modulus.setRange(1e9, 1000e9)
-        self.young_modulus.setValue(210e9)
-        self.young_modulus.setSuffix(" Pa")
-        self.young_modulus.setDecimals(2)
-        self.young_modulus.setStepType(QDoubleSpinBox.AdaptiveDecimalStepType)
-        material_layout.addRow("Young's Modulus:", self.young_modulus)
-        
-        # Density
-        self.density = QDoubleSpinBox()
-        self.density.setRange(100, 20000)
-        self.density.setValue(7800)
-        self.density.setSuffix(" kg/m³")
-        self.density.setDecimals(0)
-        material_layout.addRow("Density:", self.density)
-        
-        # Beam geometry group
-        geometry_group = QGroupBox("Beam Geometry")
-        geometry_layout = QFormLayout(geometry_group)
-        geometry_layout.setFieldGrowthPolicy(QFormLayout.AllNonFixedFieldsGrow)
-        geometry_layout.setContentsMargins(10, 20, 10, 10)
-        geometry_layout.setSpacing(10)
-        
-        # Length
-        self.beam_length = QDoubleSpinBox()
-        self.beam_length.setRange(0.1, 100)
-        self.beam_length.setValue(10.0)
-        self.beam_length.setSuffix(" m")
-        self.beam_length.setDecimals(2)
-        geometry_layout.addRow("Length:", self.beam_length)
-        
-        # Width
-        self.beam_width = QDoubleSpinBox()
-        self.beam_width.setRange(0.001, 10)
-        self.beam_width.setValue(0.2)
-        self.beam_width.setSuffix(" m")
-        self.beam_width.setDecimals(3)
-        geometry_layout.addRow("Width:", self.beam_width)
-        
-        # Height
-        self.beam_height = QDoubleSpinBox()
-        self.beam_height.setRange(0.001, 10)
-        self.beam_height.setValue(0.3)
-        self.beam_height.setSuffix(" m")
-        self.beam_height.setDecimals(3)
-        geometry_layout.addRow("Height:", self.beam_height)
-        
-        # Number of elements
-        self.num_elements = QSpinBox()
-        self.num_elements.setRange(2, 100)
-        self.num_elements.setValue(10)
-        geometry_layout.addRow("Number of Elements:", self.num_elements)
-        
-        # Spring stiffness
-        self.k_spring = QDoubleSpinBox()
-        self.k_spring.setRange(0, 1e9)
-        self.k_spring.setValue(1e5)
-        self.k_spring.setSuffix(" N/m")
-        self.k_spring.setDecimals(0)
-        self.k_spring.setStepType(QDoubleSpinBox.AdaptiveDecimalStepType)
-        geometry_layout.addRow("Tip Spring Stiffness:", self.k_spring)
-        
-        # Add groups to layout
-        beam_layout.addWidget(material_group)
-        beam_layout.addWidget(geometry_group)
-        beam_layout.addStretch()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(beam_tab, "Beam Parameters")
-        
-    def init_layers_tab(self):
-        """Initialize the layers tab for composite beam"""
-        layers_tab = QWidget()
-        main_layout = QHBoxLayout(layers_tab)
-        main_layout.setContentsMargins(20, 20, 20, 20)
-        main_layout.setSpacing(15)
-        
-        # Left side - Cross-section visualization (takes 60% of width)
-        viz_container = QWidget()
-        viz_layout = QVBoxLayout(viz_container)
-        viz_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Add title and description
-        viz_title = QLabel("Cross-Section Visualization")
-        viz_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        viz_layout.addWidget(viz_title)
-        
-        viz_desc = QLabel("Visual representation of the beam's cross-section with layers")
-        viz_desc.setStyleSheet("color: #666; font-style: italic;")
-        viz_layout.addWidget(viz_desc)
-        
-        # Cross-section visualization - make it larger
-        self.cross_section_visualizer = CrossSectionVisualizer()
-        self.cross_section_visualizer.setMinimumHeight(350)
-        viz_layout.addWidget(self.cross_section_visualizer, 1)  # stretch factor of 1
-        
-        # Add dimension information
-        dimension_label = QLabel("Total Height: 0.0 m")
-        dimension_label.setAlignment(Qt.AlignCenter)
-        self.dimension_label = dimension_label  # Store for later updates
-        viz_layout.addWidget(dimension_label)
-        
-        main_layout.addWidget(viz_container, 60)  # 60% of width
-        
-        # Right side - Layers table and controls
-        controls_container = QWidget()
-        controls_layout = QVBoxLayout(controls_container)
-        controls_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Add title
-        controls_title = QLabel("Layer Properties")
-        controls_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        controls_layout.addWidget(controls_title)
-        
-        # Layers table
-        self.layers_table = QTableWidget()
-        self.layers_table.setColumnCount(3)
-        self.layers_table.setHorizontalHeaderLabels(["Height (m)", "Young's Modulus (Pa)", "Density (kg/m³)"])
-        self.layers_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.layers_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.layers_table.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.layers_table.setMinimumHeight(200)
-        controls_layout.addWidget(self.layers_table)
-        
-        # Layer buttons
-        btn_layout = QHBoxLayout()
-        
-        self.add_layer_btn = QPushButton("Add Layer")
-        self.add_layer_btn.clicked.connect(self.add_new_layer)
-        btn_layout.addWidget(self.add_layer_btn)
-        
-        self.edit_layer_btn = QPushButton("Edit Layer")
-        self.edit_layer_btn.clicked.connect(self.edit_layer)
-        self.edit_layer_btn.setEnabled(False)
-        btn_layout.addWidget(self.edit_layer_btn)
-        
-        self.remove_layer_btn = QPushButton("Remove Layer")
-        self.remove_layer_btn.clicked.connect(self.remove_layer)
-        self.remove_layer_btn.setEnabled(False)
-        btn_layout.addWidget(self.remove_layer_btn)
-        
-        # Connect selection change
-        self.layers_table.itemSelectionChanged.connect(self.update_layer_buttons)
-        
-        # Add button layout
-        controls_layout.addLayout(btn_layout)
-        
-        # Add help text
-        help_text = QLabel(
-            "Add layers to define a composite beam. The total height "
-            "will be the sum of all layer heights. Layers are stacked "
-            "from top to bottom as shown in the visualization."
-        )
-        help_text.setWordWrap(True)
-        help_text.setStyleSheet("color: #666; font-style: italic;")
-        controls_layout.addWidget(help_text)
-        
-        # Add a stretch to push everything up
-        controls_layout.addStretch()
-        
-        main_layout.addWidget(controls_container, 40)  # 40% of width
-        
-        # Add existing layers to the table
-        for layer in self.layers:
-            self.add_layer_to_table(layer)
-        
-        # Update the cross-section
-        self.update_cross_section()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(layers_tab, "Layers")
-    
-    def update_layer_buttons(self):
-        """Enable or disable layer editing buttons based on selection"""
-        selected = len(self.layers_table.selectedItems()) > 0
-        self.edit_layer_btn.setEnabled(selected)
-        self.remove_layer_btn.setEnabled(selected)
-    
-    def add_new_layer(self):
-        """Open dialog to add a new layer"""
-        dialog = LayerDialog(self)
-        if dialog.exec_():
-            layer_data = dialog.get_layer_data()
-            
-            # Convert string expressions to callables if needed
-            for key in ['E', 'rho']:
-                if isinstance(layer_data.get(key), str):
-                    from Continues_beam.utils import parse_expression
-                    try:
-                        layer_data[key] = parse_expression(layer_data[key])
-                    except ValueError as e:
-                        QMessageBox.warning(self, "Expression Error", str(e))
-                        return
-            
-            # Add to layers list
-            self.layers.append(layer_data)
-            
-            # Add to table
-            self.add_layer_to_table(layer_data)
-            
-            # Update visualization
-            self.update_cross_section()
-    
-    def edit_layer(self):
-        """Edit the selected layer"""
-        selected_row = self.layers_table.currentRow()
-        if selected_row >= 0 and selected_row < len(self.layers):
-            # Get current layer data
-            layer_data = self.layers[selected_row]
-            
-            # Open dialog with current data
-            dialog = LayerDialog(self, layer_data)
-            if dialog.exec_():
-                new_layer_data = dialog.get_layer_data()
-                
-                # Convert string expressions to callables if needed
-                for key in ['E', 'rho']:
-                    if isinstance(new_layer_data.get(key), str):
-                        from Continues_beam.utils import parse_expression
-                        try:
-                            new_layer_data[key] = parse_expression(new_layer_data[key])
-                        except ValueError as e:
-                            QMessageBox.warning(self, "Expression Error", str(e))
-                            return
-                
-                # Update layers list
-                self.layers[selected_row] = new_layer_data
-                
-                # Update table display
-                self.layers_table.setRowCount(0)
-                for layer in self.layers:
-                    self.add_layer_to_table(layer)
-                
-                # Update visualization
-                self.update_cross_section()    
-    def remove_layer(self):
-        """Remove the selected layer"""
-        selected_row = self.layers_table.currentRow()
-        if selected_row >= 0 and selected_row < len(self.layers):
-            # Confirm deletion
-            reply = QMessageBox.question(
-                self, 
-                "Confirm Deletion",
-                f"Remove layer {selected_row + 1}?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No
-            )
-            
-            if reply == QMessageBox.Yes:
-                # Remove from layers list
-                del self.layers[selected_row]
-                
-                # Update table display
-                self.layers_table.setRowCount(0)
-                for layer in self.layers:
-                    self.add_layer_to_table(layer)
-                
-                # Update visualization
-                self.update_cross_section()
-                
-    def add_layer_to_table(self, layer):
-        """Add a layer to the table display
-        
-        Args:
-            layer (dict): Layer data dictionary
-        """
-        row = self.layers_table.rowCount()
-        self.layers_table.insertRow(row)
-        
-        # Height column
-        height_item = QTableWidgetItem(f"{layer['height']:.4f}")
-        height_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 0, height_item)
-        
-        # Young's modulus column
-        if 'E_expr' in layer:
-            E_text = layer['E_expr']
-        else:
-            E_text = f"{layer['E']:.2e}"
-        E_item = QTableWidgetItem(E_text)
-        E_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 1, E_item)
-        
-        # Density column
-        if 'rho_expr' in layer:
-            rho_text = layer['rho_expr']
-        else:
-            rho_text = f"{layer['rho']:.2f}"
-        rho_item = QTableWidgetItem(rho_text)
-        rho_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 2, rho_item)
-        
-    def init_loads_tab(self):
-        """Initialize the loads & time tab for continuous beam analysis"""
-        loads_tab = QWidget()
-        layout = QVBoxLayout(loads_tab)
-        layout.setContentsMargins(20, 20, 20, 20)
-        layout.setSpacing(15)
-        
-        # Create panel for managing force regions
-        regions_group = QGroupBox("Force Regions")
-        regions_layout = QVBoxLayout(regions_group)
-        self.force_regions_panel = ForceRegionsPanel(self.force_regions_manager)
-        regions_layout.addWidget(self.force_regions_panel)
-        layout.addWidget(regions_group)
-        
-        # Time settings group
-        time_group = QGroupBox("Time Settings")
-        time_layout = QFormLayout(time_group)
-        time_group.setLayout(time_layout)
-        
-        # Start time
-        self.time_start_spin = QDoubleSpinBox()
-        self.time_start_spin.setRange(0, 1000)
-        self.time_start_spin.setValue(0.0)
-        self.time_start_spin.setSuffix(" s")
-        self.time_start_spin.setDecimals(2)
-        time_layout.addRow("Start Time:", self.time_start_spin)
-        
-        # End time
-        self.time_end_spin = QDoubleSpinBox()
-        self.time_end_spin.setRange(0.01, 1000)
-        self.time_end_spin.setValue(3.0)
-        self.time_end_spin.setSuffix(" s")
-        self.time_end_spin.setDecimals(2)
-        time_layout.addRow("End Time:", self.time_end_spin)
-        
-        # Number of time points
-        self.time_points_spin = QSpinBox()
-        self.time_points_spin.setRange(10, 10000)
-        self.time_points_spin.setValue(300)
-        time_layout.addRow("Number of Time Points:", self.time_points_spin)
-        
-        layout.addWidget(time_group)
-        
-        # Stretch at the end to push everything to the top
-        layout.addStretch()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(loads_tab, "Loads & Time")
-        
-    def update_cross_section(self):
-        """Update the cross-section visualizer with current layers"""
-        beam_width = self.beam_width.value()
-        self.cross_section_visualizer.set_layers(self.layers, beam_width)
-        
-        # Calculate the total height of all layers
-        total_height = sum(layer.get('height', 0.0) for layer in self.layers) if self.layers else 0.0
-        
-        # Update the dimension label with the total height
-        self.dimension_label.setText(f"Total Height: {total_height:.4f} m")
-        
-        # Ensure the visualizer repaints itself
-        self.cross_section_visualizer.update()
-
-    def init_beam_results_tab(self):
-        """Initialize the results tab for continuous beam analysis"""
-        # Create the new comprehensive results dashboard
-        try:
-            # Try to import from the new location first
-            try:
-                from src.ui.components.results_dashboard import ResultsDashboard
-            except ImportError:
-                # Fall back to the old location
-                from Continues_beam.ui.results_dashboard import ResultsDashboard
-            
-            results_tab = QWidget()
-            layout = QVBoxLayout(results_tab)
-            layout.setContentsMargins(0, 0, 0, 0)
-            layout.setSpacing(0)
-            
-            # Add saving capability header
-            header_layout = QHBoxLayout()
-            header_layout.setContentsMargins(10, 5, 10, 5)
-            
-            # Add save results button in the header
-            self.save_results_btn = QPushButton("💾 Save Results")
-            self.save_results_btn.clicked.connect(self.save_beam_results)
-            self.save_results_btn.setEnabled(False)  # Disabled until results are available
-            header_layout.addStretch()
-            header_layout.addWidget(self.save_results_btn)
-            
-            layout.addLayout(header_layout)
-            
-            # Create the dashboard
-            self.results_dashboard = ResultsDashboard()
-            
-            # Set size policy to allow the dashboard to expand
-            self.results_dashboard.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-            
-            # Add the dashboard to the layout with a stretch factor
-            layout.addWidget(self.results_dashboard, 1)  # Give it a stretch factor of 1
-            
-            # For backward compatibility, maintain the frequency text widget
-            self.freq_text = QTextEdit()
-            self.freq_text.setVisible(False)  # Hide it as it's not needed in the new UI
-            
-            # Add to tabs
-            self.beam_tabs.addTab(results_tab, "Results")
-            
-        except ImportError as e:
-            # Fallback to old implementation if ResultsDashboard is not available
-            print(f"Warning: Could not import ResultsDashboard: {e}")
-            print("Using legacy results display instead.")
-            
-            results_tab = QWidget()
-            layout = QVBoxLayout(results_tab)
-            layout.setContentsMargins(5, 5, 5, 5)
-            layout.setSpacing(5)
-            
-            # Results display area
-            results_group = QGroupBox("Analysis Results")
-            results_layout = QVBoxLayout(results_group)
-            results_layout.setContentsMargins(5, 10, 5, 5)
-            results_layout.setSpacing(5)
-            
-            # Add natural frequencies display in a compact form
-            freq_header_layout = QHBoxLayout()
-            freq_label = QLabel("Natural Frequencies:")
-            freq_label.setFont(QFont("Segoe UI", 10, QFont.Bold))
-            freq_header_layout.addWidget(freq_label)
-            freq_header_layout.addStretch()
-            
-            # Add save results button in the header
-            self.save_results_btn = QPushButton("💾 Save Results")
-            self.save_results_btn.clicked.connect(self.save_beam_results)
-            self.save_results_btn.setEnabled(False)  # Disabled until results are available
-            freq_header_layout.addWidget(self.save_results_btn)
-            
-            results_layout.addLayout(freq_header_layout)
-            
-            self.freq_text = QTextEdit()
-            self.freq_text.setReadOnly(True)
-            self.freq_text.setMaximumHeight(80)  # Reduce height to save space
-            results_layout.addWidget(self.freq_text)
-            
-            # Create a splitter for the visualization area
-            viz_splitter = QSplitter(Qt.Horizontal)
-            viz_splitter.setChildrenCollapsible(False)
-            
-            # Left side - Static visualizations
-            left_viz = QTabWidget()
-            left_viz.setDocumentMode(True)
-            
-            # Beam deflection tab
-            beam_tab = QWidget()
-            beam_layout = QVBoxLayout(beam_tab)
-            beam_layout.setContentsMargins(2, 2, 2, 2)
-            self.beam_canvas = PlotCanvas(beam_tab)
-            beam_layout.addWidget(self.beam_canvas)
-            
-            # Node displacement over time tab
-            node_tab = QWidget()
-            node_layout = QVBoxLayout(node_tab)
-            node_layout.setContentsMargins(2, 2, 2, 2)
-            
-            # Node selection
-            node_select_layout = QHBoxLayout()
-            node_select_layout.addWidget(QLabel("Select Node:"))
-            
-            self.node_combo = QComboBox()
-            node_select_layout.addWidget(self.node_combo)
-            node_select_layout.addStretch()
-            
-            node_layout.addLayout(node_select_layout)
-            
-            self.node_canvas = PlotCanvas(node_tab)
-            node_layout.addWidget(self.node_canvas)
-            
-            # Add tabs to left visualization
-            left_viz.addTab(beam_tab, "Beam Deflection")
-            left_viz.addTab(node_tab, "Node Displacement")
-            
-            # Right side - Animations
-            right_viz = QTabWidget()
-            right_viz.setDocumentMode(True)
-            
-            # Beam animation tab
-            beam_anim_tab = QWidget()
-            beam_anim_layout = QVBoxLayout(beam_anim_tab)
-            beam_anim_layout.setContentsMargins(2, 2, 2, 2)
-            self.beam_animation_adapter = BeamAnimationAdapter()
-            beam_anim_layout.addWidget(self.beam_animation_adapter)
-            
-            # Mode shape animation tab
-            mode_shape_tab = QWidget()
-            mode_shape_layout = QVBoxLayout(mode_shape_tab)
-            mode_shape_layout.setContentsMargins(2, 2, 2, 2)
-            self.mode_shape_adapter = ModeShapeAdapter()
-            mode_shape_layout.addWidget(self.mode_shape_adapter)
-            
-            # Add tabs to right visualization
-            right_viz.addTab(beam_anim_tab, "Beam Animation")
-            right_viz.addTab(mode_shape_tab, "Mode Shapes")
-            
-            # Add both sides to the splitter
-            viz_splitter.addWidget(left_viz)
-            viz_splitter.addWidget(right_viz)
-            
-            # Set initial sizes to make them equal
-            viz_splitter.setSizes([500, 500])
-            
-            # Make the visualization area taller
-            viz_splitter.setMinimumHeight(1000)
-            
-            # Add the splitter to the results layout with stretch factor
-            results_layout.addWidget(viz_splitter, 1)  # Give it a stretch factor of 1
-            
-            # Add the results group to the main layout with stretch factor
-            layout.addWidget(results_group, 1)  # Give it a stretch factor of 1
-            
-            # Add to tabs
-            self.beam_tabs.addTab(results_tab, "Results")
-        
-    def run_beam_simulation(self):
-        """Run the beam vibration simulation"""
-        try:
-            # Update status
-            self.status_bar.showMessage("Running simulation...")
-            
-            # Get beam parameters
-            beam_length = self.beam_length.value()
-            beam_width = self.beam_width.value()
-            spring_constant = self.k_spring.value()
-            num_elements = self.num_elements.value()
-            
-            # Validate inputs - ensure we have at least one layer
-            if not self.layers or all(layer.get('height', 0) <= 0 for layer in self.layers):
-                QMessageBox.warning(self, "Simulation Error", 
-                                   "You must define at least one layer with a positive height.")
-                self.statusBar().showMessage("Simulation failed: No valid layers defined")
-                return
-            
-            # Prepare layers for solver
-            layers = []
-            for layer in self.layers:
-                height = layer.get('height', 0)
-                E = layer.get('E', 0)
-                rho = layer.get('rho', 0)
-                
-                layers.append({'height': height, 'E': E, 'rho': rho})
-            
-            # Set up force function
-            force_gens = get_force_generators()
-            
-            # Check if we have any defined force regions
-            if not self.force_regions_manager.regions:
-                msg = QMessageBox()
-                msg.setIcon(QMessageBox.Information)
-                msg.setWindowTitle("No Force Regions")
-                msg.setText("No force regions are defined.")
-                msg.setInformativeText("The simulation will run with a zero force function.\n"
-                                       "Consider adding force regions in the Loads & Time tab for meaningful results.")
-                msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-                response = msg.exec_()
-                
-                if response == QMessageBox.Cancel:
-                    self.statusBar().showMessage("Simulation cancelled")
-                    return
-                
-                # Create a lambda function that always returns 0.0 as the force profile
-                f_profile = lambda x, t: 0.0
-            else:
-                # Create force profile from defined regions
-                f_profile = self.force_regions_manager.create_force_function(force_gens)
-            
-            # Run simulation
-            results = solve_beam_vibration(
-                width=beam_width,
-                layers=layers,
-                L=beam_length,
-                k_spring=spring_constant,
-                num_elems=num_elements,
-                f_profile=f_profile
-            )
-            
-            # Store results
-            self.simulation_results = results
-            
-            # Update display
-            self.update_results_display()
-            
-            # Update status
-            self.status_bar.showMessage("Simulation completed successfully")
-            
-        except Exception as e:
-            error_message = str(e)
-            detailed_message = "An error occurred during the simulation."
-            
-            # Check for common errors
-            if "no object chosen" in error_message.lower():
-                detailed_message = "A reference to an undefined object was detected. "
-                detailed_message += "This may happen if no force regions are defined or if there's an issue with the layers configuration."
-            
-            # Show error message
-            msg_box = QMessageBox()
-            msg_box.setIcon(QMessageBox.Critical)
-            msg_box.setWindowTitle("Simulation Error")
-            msg_box.setText(detailed_message)
-            msg_box.setDetailedText(f"Error details: {error_message}")
-            msg_box.exec_()
-            
-            # Update status bar
-            self.status_bar.showMessage(f"Simulation failed: {error_message}")
-    
-    def update_results_display(self):
-        """Update the results display with simulation results"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if we're using the new results dashboard
-        if hasattr(self, 'results_dashboard'):
-            # Update the comprehensive dashboard with all results
-            self.results_dashboard.update_results(self.simulation_results)
-            
-            # Ensure the freq_text is also populated for backward compatibility
-            self.freq_text.clear()
-            for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                if i < 10:  # Show only first 10 modes
-                    self.freq_text.append(f"Mode {i+1}: {freq:.2f} Hz")
-        else:
-            # Use the legacy display method
-            # Display natural frequencies
-            self.freq_text.clear()
-            for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                if i < 10:  # Show only first 10 modes
-                    self.freq_text.append(f"Mode {i+1}: {freq:.2f} Hz")
-            
-            # Update beam deflection plot
-            self.plot_beam_deflection()
-            
-            # Update node selection combo
-            self.node_combo.clear()
-            for i in range(len(self.simulation_results['coords'])):
-                x = self.simulation_results['coords'][i]
-                self.node_combo.addItem(f"Node {i+1} (x={x:.2f}m)")
-            
-            # Connect node selection to plot update
-            self.node_combo.currentIndexChanged.connect(self.plot_node_displacement)
-            
-            # Initial node plot
-            if self.node_combo.count() > 0:
-                self.node_combo.setCurrentIndex(0)
-                self.plot_node_displacement()
-                
-            # Update animations
-            if hasattr(self, 'beam_animation_adapter'):
-                self.beam_animation_adapter.update_animation(self.simulation_results)
-                
-            if hasattr(self, 'mode_shape_adapter'):
-                self.mode_shape_adapter.update_results(self.simulation_results)
-            
-        # Enable save results button
-        self.save_results_btn.setEnabled(True)
-        
-    def plot_beam_deflection(self):
-        """Plot the beam deflection"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if canvas exists
-        if not hasattr(self, 'beam_canvas') or self.beam_canvas is None:
-            return
-        
-        # Clear the canvas
-        self.beam_canvas.clear()
-        ax = self.beam_canvas.figure.add_subplot(111)
-        
-        # Get data
-        x = self.simulation_results['coords']
-        t = self.simulation_results['time']
-        u = self.simulation_results['displacement']
-        
-        # Plot beam deflection at selected time points
-        num_frames = min(10, len(t))
-        step = len(t) // num_frames
-        
-        # Extract displacement at even nodes (translations)
-        u_disp = u[::2, :]
-        
-        for i in range(0, len(t), step):
-            if i >= len(t):
-                continue
-            
-            # Get displacement at this time
-            deflection = u_disp[:, i]
-            
-            # Scale factor for visualization
-            scale = 1.0
-            if np.max(np.abs(deflection)) > 0:
-                scale = 0.2 * np.max(x) / np.max(np.abs(deflection))
-            
-            # Plot the deflected beam
-            ax.plot(x, scale * deflection, label=f"t={t[i]:.2f}s")
-        
-        # Plot settings
-        ax.set_xlabel('Position (m)')
-        ax.set_ylabel('Deflection (scaled)')
-        ax.set_title('Beam Deflection')
-        ax.grid(True)
-        ax.legend()
-        
-        # Update canvas
-        self.beam_canvas.draw()
-        
-    def plot_node_displacement(self):
-        """Plot the displacement of the selected node over time"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if canvas exists
-        if not hasattr(self, 'node_canvas') or self.node_canvas is None:
-            return
-        
-        # Get selected node
-        if self.node_combo.count() == 0:
-            return
-        
-        node_idx = self.node_combo.currentIndex()
-        
-        # Clear the canvas
-        self.node_canvas.clear()
-        ax = self.node_canvas.figure.add_subplot(111)
-        
-        # Get data
-        t = self.simulation_results['time']
-        u = self.simulation_results['displacement']
-        v = self.simulation_results['velocity']
-        a = self.simulation_results['acceleration']
-        
-        # Get the displacement DOF for this node
-        dof = node_idx * 2  # 2 DOFs per node (displacement and rotation)
-        
-        # Plot displacement, velocity, and acceleration
-        ax.plot(t, u[dof, :], label='Displacement')
-        ax.plot(t, v[dof, :], label='Velocity')
-        ax.plot(t, a[dof, :] / 1000, label='Acceleration (÷1000)')
-        
-        # Plot settings
-        ax.set_xlabel('Time (s)')
-        ax.set_ylabel('Amplitude')
-        ax.set_title(f'Node {node_idx+1} Response')
-        ax.grid(True)
-        ax.legend()
-        
-        # Update canvas
-        self.node_canvas.draw()
-        
-    def save_beam_results(self):
-        """Save beam simulation results to CSV files"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            QMessageBox.warning(self, "No Results", "Please run a simulation before saving results.")
-            return
-            
-        try:
-            # Ask for directory to save results
-            save_dir = QFileDialog.getExistingDirectory(self, "Select Directory to Save Results")
-            if not save_dir:
-                return
-                
-            # Save natural frequencies
-            freq_file = os.path.join(save_dir, "natural_frequencies.csv")
-            with open(freq_file, 'w') as f:
-                f.write("Mode,Frequency (Hz)\n")
-                for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                    f.write(f"{i+1},{freq:.6f}\n")
-                    
-            # Save displacement results
-            time = self.simulation_results['time']
-            disp = self.simulation_results['displacement']
-            coords = self.simulation_results['coords']
-            
-            # Create a DataFrame for displacement over time
-            disp_data = pd.DataFrame()
-            disp_data['Time (s)'] = time
-            
-            # Add columns for each node's displacement
-            for i in range(len(coords)):
-                dof = i * 2  # 2 DOFs per node (displacement and rotation)
-                disp_data[f'Node {i+1} (x={coords[i]:.2f}m)'] = disp[dof, :]
-                
-            # Save to CSV
-            disp_file = os.path.join(save_dir, "displacement_results.csv")
-            disp_data.to_csv(disp_file, index=False)
-            
-            # Create a DataFrame for the currently selected node's detailed response
-            if self.node_combo.count() > 0:
-                node_idx = self.node_combo.currentIndex()
-                dof = node_idx * 2
-                
-                node_data = pd.DataFrame()
-                node_data['Time (s)'] = time
-                node_data['Displacement'] = self.simulation_results['displacement'][dof, :]
-                node_data['Velocity'] = self.simulation_results['velocity'][dof, :]
-                node_data['Acceleration'] = self.simulation_results['acceleration'][dof, :]
-                
-                # Save to CSV
-                node_file = os.path.join(save_dir, f"node_{node_idx+1}_response.csv")
-                node_data.to_csv(node_file, index=False)
-            
-            # Show success message
-            QMessageBox.information(self, "Results Saved", 
-                                   f"Results successfully saved to {save_dir}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Error Saving Results", f"Error: {str(e)}")
-
-    def format_parameter_name(self, param):
-        """
-        Format parameter names using LaTeX symbols for plotting.
-
-        Parameters:
-            param (str): The parameter name.
-
-        Returns:
-            str: Formatted parameter name.
-        """
-        GREEK_LETTERS = {
-            'beta': r'\beta',
-            'lambda': r'\lambda',
-            'mu': r'\mu',
-            'nu': r'\nu'
-        }
-        for greek_letter, symbol in GREEK_LETTERS.items():
-            if param.startswith(greek_letter):
-                index = param[len(greek_letter) + 1:]  # Capture the index after _
-                return f'${symbol}_{{{index}}}$'
-        return param.replace("_", " ")
-
-    def create_comparative_plot(self):
-        """Create a comparative plot from selected FRF inputs with custom legends and title"""
-        # Make sure frf_plots is initialized
-        if not hasattr(self, 'frf_plots'):
-            self.frf_plots = {}
-            
-        # Check if we have any FRF data
-        if not self.available_plots_list or self.available_plots_list.count() == 0 or not self.frf_plots:
-            QMessageBox.warning(self, "No Data", "Please run FRF analysis first to generate plots for comparison.")
-            return
-            
-        # Get selected plots
-        selected_items = self.available_plots_list.selectedItems()
-        selected_plot_names = [item.text() for item in selected_items]
-        
-        if not selected_plot_names:
-            QMessageBox.warning(self, "Selection Error", "Please select at least one plot to compare.")
-            return
-            
-        # Check if any of the selected plots exist in frf_plots
-        valid_plot_names = [name for name in selected_plot_names if name in self.frf_plots]
-        if not valid_plot_names:
-            QMessageBox.warning(self, "Data Error", "None of the selected plots have valid FRF data.")
-            return
-        
-        # Update legend map from table
-        self.legend_map = {}
-        for row in range(self.legend_table.rowCount()):
-            original_name = self.legend_table.item(row, 0).text()
-            custom_legend = self.legend_table.item(row, 1).text() if self.legend_table.item(row, 1) else ""
-            
-            # Get style options if available
-            line_style = "-"
-            marker = ""
-            color = None
-            
-            # Get style widgets if they exist
-            line_style_widget = self.legend_table.cellWidget(row, 2)
-            marker_widget = self.legend_table.cellWidget(row, 3)
-            color_widget = self.legend_table.cellWidget(row, 4)
-            
-            if line_style_widget:
-                line_style = line_style_widget.currentText()
-                if line_style == "None":
-                    line_style = ""
-                    
-            if marker_widget:
-                marker = marker_widget.currentText()
-                if marker == "None":
-                    marker = ""
-                    
-            if color_widget:
-                # Get color from the button's background color
-                # The QPushButton doesn't have currentText() method
-                style = color_widget.styleSheet()
-                # Extract color from background-color: X; in the style sheet
-                import re
-                color_match = re.search(r'background-color:\s*(.*?);', style)
-                if color_match:
-                    color = color_match.group(1)
-                else:
-                    # Use default color if can't extract from style
-                    import random
-                    r, g, b = [random.randint(0, 255) for _ in range(3)]
-                    color = f"rgb({r},{g},{b})"
-            
-            # Store in legend map
-            self.legend_map[original_name] = {
-                'legend': custom_legend,
-                'line_style': line_style,
-                'marker': marker,
-                'color': color
-            }
-        
-        # Get custom title
-        custom_title = self.plot_title_edit.text()
-        if not custom_title:
-            custom_title = "Comparative FRF Visualization"
-        
-        # Check if figure exists, create it if not
-        if not hasattr(self, 'comp_fig') or self.comp_fig is None:
-            # Create matplotlib figure
-            from matplotlib.figure import Figure
-            from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-            
-            # Create the figure
-            self.comp_fig = Figure(figsize=(self.fig_width_spin.value(), self.fig_height_spin.value()))
-            self.comp_canvas = FigureCanvas(self.comp_fig)
-            
-            # Check if we have a layout to add it to
-            if hasattr(self, 'comp_plot_layout') and self.comp_plot_layout is not None:
-                # Remove any existing widgets
-                while self.comp_plot_layout.count():
-                    item = self.comp_plot_layout.takeAt(0)
-                    if item.widget():
-                        item.widget().deleteLater()
-                
-                # Add the new canvas
-                self.comp_plot_layout.addWidget(self.comp_canvas)
-        
-        # Clear previous figure and set its size
-        self.comp_fig.clear()
-        self.comp_fig.set_size_inches(self.fig_width_spin.value(), self.fig_height_spin.value())
-        ax = self.comp_fig.add_subplot(111)
-        
-        # Get normalization values - set to meaningful defaults and print for debugging
-        x_norm_factor = self.x_norm_value.value() if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() else 1.0
-        y_norm_factor = self.y_norm_value.value() if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() else 1.0
-        # Debug the normalization factors
-        print(f"\nDEBUG - Normalization factors: X={x_norm_factor}, Y={y_norm_factor}")
-        print(f"Normalization checkboxes: X={hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked()}, Y={hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked()}\n")
-        
-        # Plot each selected FRF
-        for plot_name in selected_plot_names:
-            if plot_name in self.frf_plots:
-                # Extract data from original plot
-                orig_fig = self.frf_plots[plot_name]
-                orig_ax = orig_fig.axes[0]
-                
-                # Extract lines data
-                for line in orig_ax.get_lines():
-                    # Get the data
-                    x_data = line.get_xdata()
-                    y_data = line.get_ydata()
-                    source_line_label = line.get_label() # Original label from the matplotlib line object
-                    # Debug the original data
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() or hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked():
-                        self.debug_array(x_data, "x_data BEFORE")
-                        self.debug_array(y_data, "y_data BEFORE")
-
-                    # Ensure we're working with numpy arrays
-                    x_data_np = np.array(x_data)
-                    y_data_np = np.array(y_data)
-
-                    # Apply normalization explicitly with numpy
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() and x_norm_factor != 1.0:
-                        x_data_np = x_data_np / x_norm_factor
-                        print(f"Applied X normalization: divided by {x_norm_factor}")
-
-                    if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() and y_norm_factor != 1.0:
-                        y_data_np = y_data_np / y_norm_factor
-                        print(f"Applied Y normalization: divided by {y_norm_factor}")
-
-                    # Always use the numpy arrays for plotting
-                    x_data = x_data_np
-                    y_data = y_data_np
-
-                    # Debug the normalized data
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() or hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked():
-                        self.debug_array(x_data, "x_data AFTER")
-                        self.debug_array(y_data, "y_data AFTER")
-
-                    # Start with defaults from the original line's properties
-                    current_legend_text = source_line_label
-                    current_line_style = line.get_linestyle()
-                    current_marker = line.get_marker()
-                    current_color = line.get_color() # Matplotlib's auto-assigned color from its cycle
-
-                    # Determine which properties from self.legend_map to use.
-                    # self.legend_map is keyed by names from column 0 of the legend_table,
-                    # which can be either a plot_name (e.g., "Mass 1 (With DVA) - RunX")
-                    # or a source_line_label (e.g., "mass_1").
-                    # We prioritize the specific source_line_label's settings if they exist.
-                    
-                    props_to_use = None
-                    if source_line_label and source_line_label in self.legend_map:
-                        # A specific entry exists for this line's original label
-                        props_to_use = self.legend_map[source_line_label]
-                    elif plot_name in self.legend_map:
-                        # No specific entry for the line's label, fall back to the overall plot_name's entry
-                        props_to_use = self.legend_map[plot_name]
-
-                    if props_to_use:
-                        # Apply custom legend text if 'legend' key explicitly exists in the properties.
-                        # This ensures that an empty string provided by the user is respected.
-                        if 'legend' in props_to_use:
-                            current_legend_text = props_to_use['legend']
-                        
-                        # Apply custom styles, falling back to current (original line) style if not specified in map.
-                        current_line_style = props_to_use.get('line_style', current_line_style)
-                        current_marker = props_to_use.get('marker', current_marker)
-                        
-                        # Get color and convert to matplotlib format if needed
-                        color_value = props_to_use.get('color', current_color)
-                        if isinstance(color_value, str) and color_value.startswith('rgb('):
-                            # Convert 'rgb(r,g,b)' format to a tuple
-                            try:
-                                import re
-                                r, g, b = map(int, re.search(r'rgb\((\d+),(\d+),(\d+)\)', color_value).groups())
-                                current_color = (r/255, g/255, b/255)  # Normalize to 0-1 range
-                            except (ValueError, AttributeError):
-                                # If conversion fails, use the original color
-                                pass
-                        else:
-                            current_color = color_value
-
-                    # Convert "None" string values from comboboxes to what Matplotlib expects
-                    if current_line_style == "None": current_line_style = "" # Empty string for no line
-                    if current_marker == "None": current_marker = "none" # "none" for no marker
-                    
-                    # Plot the line on the comparative figure with customized style
-                    ax.plot(x_data, y_data, label=current_legend_text,
-                            linestyle=current_line_style,
-                            marker=current_marker,
-                            color=current_color) # Matplotlib handles color strings and its default cycle
-        
-        # Set title with custom font size
-        ax.set_title(custom_title, fontsize=self.title_font_size.value())
-        
-        # Set axis labels with normalization factor information
-        x_label = 'Frequency (rad/s)'
-        if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() and x_norm_factor != 1.0:
-            x_label += f' / {x_norm_factor}'
-        ax.set_xlabel(x_label)
-
-        y_label = 'Amplitude'
-        if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() and y_norm_factor != 1.0:
-            y_label += f' / {y_norm_factor}'
-        ax.set_ylabel(y_label)
-        
-        # Add grid based on checkbox state
-        ax.grid(self.show_grid_check.isChecked(), linestyle='--', linewidth=0.5, alpha=0.7)
-        
-        # Add legend with custom position
-        legend_position = self.legend_position_combo.currentText()
-        ax.legend(loc=legend_position)
-        
-        # Update the figure
-        self.comp_fig.tight_layout()
-        self.comp_canvas.draw()
-        
-        # Update the legend table
-        self._update_legend_table_from_selection()
-        
-        # Show success message
-        self.status_bar.showMessage("Comparative plot created successfully")
-
-
-    def create_catalogue_creator_page(self):
-        """Empty placeholder for removed catalogue creator page"""
-        # Create empty page
-        empty_page = QWidget()
-        self.content_stack.addWidget(empty_page)
